Do not clip lines with thickness 0, ticket #53
In Postscript output, the painting of objects that can have arrowheads,
mainly polylines, is done by first defining a clipping area such that
the stroked path does not protrude beyond the arrowhead. After that, if
applicable, the path is filled, then stroked, finally the graphics state
without clipping is restored and the arrowheads are drawn. For filled
objects with line-thickness zero the clipping is not necessary, but
would lead to funny holes in the filling.