For pstricks, avoid too large numbers, ticket #87
The pstricks driver writes the instructions for drawing line objects to a string
buffer. For ridiculously large numbers, this buffer might overflow. Check the
length of the string to be written, and exit with error if the string becomes
too large.

The humble thing to do would be to malloc() a large enough buffer and cope with
whatever numbers are fed to fig2dev. Since numbers having more than one hundred
digits are probably only of interest to input fuzzers, choose the quick route
and exit.