Accept an invalid arrow type, #89, negative arrow widths
The fig format provides 29 different arrow types. These are chosen with a type
parameter set to a value from 0 to 14, and a style parameter set to either 0 or
1. The arrow with type 0, style 1 does not exist. Using that arrow lead to an
out-of-bounds array access, see ticket #89. Now, for these parameters provide
the definition of the default arrow (type 0, style 0). In addition, allow
negative arrow widths. This might be useful for asymmetric arrow tips, which
then appear mirrored at the corresponding line.