Use imagemagick or graphicsmagick if netpbm is not available
For output to bitmap formats, require ghostscript and, for most bitmap formats,
one out of either netpbm, imagemagick or graphicsmagick. The availability of
these latter programs is detected at run-time, the availability of ghostscript
at compile-time. Netpbm ist still required for output to autocad sld format.
Also, netpbm is required to read fig files which have gif, tiff or ppm images
embedded.

The test suite does not extensively test the generation of bitmap formats.
However, as a side effect of the tests on various images embedded in fig files,
the generation of some bitmap file formats is tested. In environments where all
of the above programs are missing, the tests are gracefully skipped and should
not fail.