Compile conditionally, if ghostscript is not available
Fig2dec can be built without ghostscript support. If fig2dev is used in such a
way that ghostscript is needed and (i) it was built without ghostscript support,
fig2dev fails and issues a message explaining the issue. The previous behavior
remains, (ii) if fig2dev was built with ghostscript support, but ghostscript is
not found at runtime, fig2dev fails with a less telling error message.