Pass "make check" without calling "make" before
This is of dubious quality: If make check is called before make, most tests
fail, if fig2dev is not found. If it is found, it must be an already installed
version, since fig2dev was not yet made. Now, the reason for running the test
suite is probably not to check some older version, but if you wish,
it should work.