In tests, write tiff output to a regular file
Tests partially wrote output from fig2dev to stdout, i.e., a pipe. For tiff
output, this created a specific difficulty: Fig2dev internally tries to call
pnmtotiff to create a tiff. Fig2dev also knows that pnmtotiff comes in two
versions. One older, usually installed under debian, which cannot write to a
pipe. A newer version internally creates a temporary file and can write to a
pipe. However, the test suite does not make this distinction. With this commit,
it does not need to, because tiff output anyhow goes to a regular file.