WIP: For postscript, use zlib to deflate image data
This introduces a new dependency of fig2dev on the zlib library (-lz)
and, for compilation, on the zlib header files.

Use the PostScript image operator in its form with an image dictionary
argument. Write the deflated and ASCII85 encoded data to the postcript
output. Write an image dictionary which sets the /DataSource to
"currentfile /ASCII85Decode filter" and "/FlateDecode filter".
Previously, data for images without a colormap was written uncompressed,
hex-encoded to the postscript output.

This is work-in-progress, since postscript expects RGB triplets, but the
former code works with BGR triplets. Hence, currently embedded images
without colormap have red and blue swapped.