Set a transparent color for pngs with transparency
This works fine for png images that have either fully opaque or fully
transparent pixels. Png images with a full alpha channel are first
composited against a background, either the background given in file, a
color supplied by the user (-g option), or white. The background color is
then set to be the transparency marker. This use (abuse?) of the -g option
is not documented.