Reject huge values for spline control points, #127
This is the completion of commit [e3cee2]. With commit e3cee2, the very
first control point in a spline was not checked for having a huge value.
If an output language does not natively support Bézier splines, fig2dev
translates splines to line segments. With huge values for the control
points, the translation causes a stack overflow.