Pass output language name to driver, ticket #139
The output driver is first guessed from the output filename
suffix, and the output driver is called with the -L option. If the -L
option is specified on the command line, the output driver is
called a second time, overriding the first call. In the first
call, the language name was not passed to the driver. This is now
fixed. A regression from commit [b4a152].