Pict2e output: Arcs ending at multiples of pi/4, #145, #148
For the pict2e output language, arcs are drawn by first
constructing a circle of the same radius, generated from a spline.
A section corresponding to the arc angle is then cut out from the
line forming the circle. Find the end point on the circle by
comparing the short coordinate of each point with the sinus or
cosinus, respectively, of the end angle times the radius of the
circle. Comparing the long coordinate, apart from being a glancing
intersection, may fail, since the constructed spline slightly
wobbles around the true radius. Less than 1.5 promille, but enough
to trigger an error, see tickets #145 and #148.