Accept utf8 encoded text for PostScript output
Some characters, namely those that are present in the fonts
supplied with ghostscript, can now be accessed by their
utf8-encoded unicode point. Ghostscript ships with some free fonts
for substitution of the standard 34 PostScript fonts.

The utf8 code is written as is to the show operator of postscript.
The characters in the font are accessed by using the composefont
operator to combine a hand-crafted CMap file, which maps
utf8-codes to the adobe glyph names, with the original font.

The resulting postscript file is not 7bit-clean. It can be
processed by ghostscript and was also tested with some online
converters. However, I doubt whether printers could process these
files. For printing or publishing purposes, eps or ps output from
fig2dev should be converted with eps2eps or ps2ps
(gs -sDEVICE=eps2write or gs -sDEVICE=ps2write, respectively) to
more standard postscript code.

Currently, the input fig file must be utf8-encoded.