ps-output: let 8bit-encoded symbol and dingbats text as is
Text written with the Symbol or Zapf Dingbats font may contain
characters in the range from 128 to 255. If the fig-file is utf-8
encoded, the string should be converted to latin1. Otherwise, if
for other fonts a conversion to utf-8 is needed, do not try a
conversion for Symbol or Dingbats text objects. In this way, for
8bit encodings, e.g., latin2, the characters are accessed in their
original position. Other two-byte encodings are, I believe, not
able to encode characters in the range from 128 to 255, hence a
conversion is also in this case not necessary.