Rewrite conversion of text strings to utf-8
Add a second conversion function, and prepare to use that second
conversion function. One conversion is necessary when reading legacy
fig-files, to convert text strings contained in these files from a
locale (non-utf8) encoding to utf8. The second conversion is necessary
if the current locale is in a non-utf8 encoding, and a (modern) utf-8
encoded fig-file refers to an embedded image. The file name for this
image may have to be converted to the locale encoding.